+++ src/TwigComponent/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.33
 
-- Extended support for the `index.html.twig` template fallback when resolving namespaced anonymous components.
+- Extended support for the `index.html.twig` template fallback when resolving namespaced anonymous components
+- Add support for dynamic template resolution with `AsTwigComponent(template: FromMethod('getCustomFuntion'))`
 
 ## 2.32